refactor: remove redundant shape filename suffix
