+++ browser-extensions/chrome/src/ai-spacing/content-script.ts
@@ -3,8 +3,8 @@ import type { CandidateLabel, ClassifyCandidatesMessage, ClassifyCandidatesRespo
 import { readSentence } from './sentence-context';
 import type { AmbiguousShape, SettledCandidate, TextEdit } from './shapes/base';
 import { applyTextEdits } from './shapes/base';
-import { digitPlus } from './shapes/digit-plus-shape';
-import { hyphenDigit } from './shapes/hyphen-digit-shape';
+import { digitPlus } from './shapes/digit-plus';
+import { hyphenDigit } from './shapes/hyphen-digit';
 
 const AMBIGUOUS_SHAPES: AmbiguousShape[] = [hyphenDigit, digitPlus];
